# Optimize _chunked_even_finite()