### Issue reference

#698

### Changes

- Optimize **_chunked_even_finite** by yielding the chunk as a slice of the `iterable`, without using a for-loop to build up the chunk.

### Implementation Details

`yield list(islice(...))` was used to ensure that chunks for non-list iterables such as strings are yielded as lists, and to handle non-sliceable iterables like sets and dicts.

I used the following in IPython as a rudimentary benchmark.

**OS:** Linux x64 **CPU:** AMD Ryzen 7 5800X

```python
%timeit _ = [list(chunked_even(range(x), y)) for x in range(500) for y in range(1,500)]
# Before: 2.9 s ± 24.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# After: 1.39 s ± 12.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%timeit _ = [list(chunked_even([z for z in range (x)], y)) for x in range(500) for y in range(1,500)]
# Before: 3.61 s ± 11.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# After: 1.91 s ± 19.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%timeit _ = [list(chunked_even({z:z for z in range(x)}, y)) for x in range(500) for y in range(1,500)]
# Before: 4.33 s ± 34.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# After: 2.84 s ± 24.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%timeit _ = [list(chunked_even(set(z for z in range(x)), y)) for x in range(500) for y in range(1,500)]
# Before: 4.31 s ± 26.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# After: 2.72 s ± 7.32 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```